# internal/api/dto/user_dto.go
package dto

import (
	"errors"
	"strings"
	"time"
)

type CreateUserRequest struct {
	TenantID    string `json:"tenant_id"`
	Email       string `json:"email"`
	DisplayName string `json:"display_name"`
}

func (r *CreateUserRequest) Validate() error {
	if strings.TrimSpace(r.DisplayName) == "" {
		return errors.New("display_name is required")
	}
	if strings.TrimSpace(r.Email) == "" {
		return errors.New("email is required")
	}
	// TenantID es opcional en el sistema (uuid.Nil), por lo que permitimos string vacío.
	return nil
}

type UserResponse struct {
	ID          string    `json:"id"`
	TenantID    string    `json:"tenant_id"`
	DisplayName string    `json:"display_name"`
	CreatedAt   time.Time `json:"created_at"`
}

type UpdateStatusRequest struct {
	IsActive bool `json:"is_active"`
}

func (r *UpdateStatusRequest) Validate() error {
	// Booleano no requiere validación de "required" ya que false es un valor válido.
	return nil
}

type ResetPasswordRequest struct {
	NewPassword string `json:"new_password"`
}

func (r *ResetPasswordRequest) Validate() error {
	if len(r.NewPassword) < 8 {
		return errors.New("password must be at least 8 characters")
	}
	return nil
}